refactor: DummyTerminalStation station 패키지로 이동
