+++ src/main/java/subway/line/domain/UpstreamTerminalSection.java
@@ -1,5 +1,6 @@
 package subway.line.domain;
 
+import subway.station.domain.DummyTerminalStation;
 import subway.station.domain.Station;
 
 import java.util.ArrayList;
