+++ src/main/java/subway/station/application/StationService.java
@@ -3,9 +3,12 @@ package subway.station.application;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.domain.DummyTerminalStation;
 import subway.station.domain.Station;
 import subway.station.exception.StationNotFoundException;
 import subway.station.repository.StationRepository;
+
+import java.util.Objects;
 
 @Transactional
 @Service
@@ -20,6 +23,9 @@ public class StationService {
 
     @Transactional(readOnly = true)
     public Station findStationByName(String stationName) {
+        if (Objects.equals(stationName, DummyTerminalStation.STATION_NAME)) {
+            return DummyTerminalStation.getInstance();
+        }
         return stationRepository.findStationByName(stationName)
                                 .orElseThrow(() -> new StationNotFoundException("존재하지 않는 역 이름입니다."));
     }
