+++ src/test/java/subway/domain/LineTest.java
@@ -4,12 +4,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import subway.line.domain.DummyTerminalStation;
 import subway.line.domain.Line;
 import subway.line.domain.MiddleSection;
 import subway.line.exception.DuplicateStationInLineException;
 import subway.line.exception.InvalidAdditionalFareException;
 import subway.line.exception.SectionNotFoundException;
+import subway.station.domain.DummyTerminalStation;
 import subway.station.domain.Station;
 import subway.station.exception.NameLengthException;
 import subway.station.exception.StationNotFoundException;
