+++ src/test/java/subway/domain/MiddleSectionTest.java
@@ -4,10 +4,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import subway.line.domain.DummyTerminalStation;
 import subway.line.domain.MiddleSection;
 import subway.line.exception.InvalidDistanceException;
 import subway.line.exception.InvalidStationNameException;
+import subway.station.domain.DummyTerminalStation;
 import subway.station.domain.Station;
 
 import static org.assertj.core.api.Assertions.*;
